+++ stdlib/rust/src/lib.rs
@@ -1,33 +1,18 @@
-#![no_std]
 
 extern crate alloc;
+extern crate core;
+
 mod wasome_mem;
 mod wasome_string;
 mod wasome_vec;
+mod plattform;
 
 use crate::wasome_string::WasomeString;
 use alloc::boxed::Box;
 use alloc::string::String;
 use core::mem::forget;
-use core::panic::PanicInfo;
+use crate::plattform::{exit, print, read_line_internal};
 
-#[link(wasm_import_module = "wasi_snapshot_preview1")]
-unsafe extern "C" {
-    fn fd_write(fd: i32, iovs: *const Iovec, iovs_len: usize, nwritten: *mut usize) -> i32;
-    fn fd_read(fd: i32, iovs: *const MutIovec, iovs_len: usize, nwritten: *mut usize) -> i32;
-    safe fn proc_exit(code: i32);
-}
-#[repr(C)]
-struct Iovec {
-    buf: *const u8,
-    len: usize,
-}
-
-#[repr(C)]
-struct MutIovec {
-    buf: *mut u8,
-    len: usize,
-}
 #[global_allocator]
 static A: dlmalloc::GlobalDlmalloc = dlmalloc::GlobalDlmalloc;
 /// # Safety
@@ -69,66 +54,11 @@ impl WasomeComposite {
     }
 }
 
-fn print(s: &str) {
-    let iovec = Iovec {
-        buf: s.as_ptr(),
-        len: s.len(),
-    };
-    let mut nwritten: usize = 0;
-    // SAFETY:
-    // FD 1 always exists
-    unsafe {
-        fd_write(1, &iovec, 1, &mut nwritten);
-    }
-}
-
-fn read_line_internal() -> String {
-    let mut res = String::new();
-    let mut curr_char: u32 = 0;
-    loop {
-        let mut buf = [0_u8; 1];
-        let iovec = MutIovec {
-            buf: buf.as_mut_ptr(),
-            len: 1,
-        };
-        let mut nwritten: usize = 0;
-        // SAFETY:
-        // FD 0 always exists
-        unsafe {
-            fd_read(0, &iovec, 1, &mut nwritten);
-        }
-        if nwritten != 1 {
-            panic!("Read failed! {nwritten}");
-        }
-        curr_char <<= 8;
-        curr_char |= buf[0] as u32;
-        if let Some(converted) = char::from_u32(curr_char) {
-            if curr_char == b'\n' as u32 {
-                break;
-            }
-            res.push(converted);
-            curr_char = 0;
-        }
-    }
-    res
-}
-
 #[unsafe(no_mangle)]
 pub extern "C" fn read_line() -> *mut WasomeString {
     let wasome_string = Box::into_raw(Box::new(WasomeString::new()));
     unsafe { WasomeString::update_from_string(wasome_string, read_line_internal()) }
     wasome_string
-}
-
-fn exit(code: i32) {
-    proc_exit(code)
-}
-
-// Prevent errors
-#[cfg(not(test))]
-#[panic_handler]
-fn rust_panic(info: &PanicInfo) -> ! {
-    panic_internal(info.message().as_str())
 }
 
 fn panic_internal(msg: Option<&str>) -> ! {
